refactor: rename api package in libraries
