+++ scafi-mp-api/js-native/src/main/scala/it/unibo/scafi/libraries/PortableFieldBasedSharedData.scala
@@ -1,4 +1,4 @@
-package it.unibo.scafi.api
+package it.unibo.scafi.libraries
 
 /**
  * A portable definition of a field-based `SharedData` structure, useful for non-jvm platforms.
