+++ scafi-mp-api/js/src/main/scala/it/unibo/scafi/libraries/FullLibrary.scala
@@ -1,4 +1,4 @@
-package it.unibo.scafi.api
+package it.unibo.scafi.libraries
 
 import scala.scalajs.js.annotation.JSExportTopLevel
