+++ scafi-mp-api/js/src/test/scala/it/unibo/scafi/libraries/FullLibraryTest.scala
@@ -1,4 +1,4 @@
-package it.unibo.scafi.api
+package it.unibo.scafi.libraries
 
 class FullLibraryTest extends JSLibraryTest:
