+++ scafi-mp-api/shared/src/main/scala/it/unibo/scafi/libraries/FullPortableLibrary.scala
@@ -1,4 +1,4 @@
-package it.unibo.scafi.api
+package it.unibo.scafi.libraries
 
 import it.unibo.scafi.language.AggregateFoundation
 import it.unibo.scafi.language.common.syntax.BranchingSyntax
