+++ scafi-mp-api/shared/src/main/scala/it/unibo/scafi/libraries/PortableBranchingLibrary.scala
@@ -1,4 +1,4 @@
-package it.unibo.scafi.api
+package it.unibo.scafi.libraries
 
 /**
  * The portable library providing domain branching syntax.
