+++ scafi-mp-api/shared/src/main/scala/it/unibo/scafi/libraries/PortableExchangeCalculusLibrary.scala
@@ -1,4 +1,4 @@
-package it.unibo.scafi.api
+package it.unibo.scafi.libraries
 
 import scala.util.chaining.scalaUtilChainingOps
