+++ scafi-mp-api/shared/src/main/scala/it/unibo/scafi/libraries/ReturnSending.scala
@@ -1,4 +1,4 @@
-package it.unibo.scafi.api
+package it.unibo.scafi.libraries
 
 import scala.scalajs.js.annotation.{ JSExport, JSExportTopLevel }
