+++ scafi-mp-api/shared/src/main/scala/it/unibo/scafi/utils/libraries/IsoUtils.scala
@@ -1,4 +1,4 @@
-package it.unibo.scafi.utils.api
+package it.unibo.scafi.utils.libraries
 
 import monocle.Iso
